+++ Assets/Scripts/Game/Unit/AreaExecution/IceAreaEffectExecution.cs
@@ -14,7 +14,7 @@ public class IceAreaEffectExecution : RetangleAreaEffectExecution
         resourcePath += "IceAreaEffect";
     }
 
-    public void Init(BaseUnit creator, int time, int currentRowIndex, int colCount, int rowCount, float offsetX, int offsetY, bool isAffectFood, bool isAffectMouse)
+    public void Init(BaseUnit creator, int time, int currentRowIndex, float colCount, int rowCount, float offsetX, int offsetY, bool isAffectFood, bool isAffectMouse)
     {
         this.creator = creator;
         this.time = time;
